requirements.txt
Django>=5.0
mysqlclient
python-dotenv
requests
google-generativeai
